pyproject.toml
[project]
name = "slack-bot-daily-summary"
version = "0.1.0"
description = ""
authors = [
    {name = "JohnSmith",email = "me@flavienleblond.com"}
]
readme = "README.md"
requires-python = "<4.0,>=3.13"
dependencies = [
    "slack-sdk (>=3.36.0,<4.0.0)",
    "schedule (>=1.2.2,<2.0.0)",
    "openai (>=1.99.6,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "tinydb (>=4.8.2,<5.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
]

[project.optional-dependencies]
dev = [
    "ty>=0.0.1a19",
    "ruff>=0.12.10",
    "responses (>=0.25.8,<0.26.0)",
    "pytest-cov (>=6.2.1,<7.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
]

[tool.ty.src]
exclude = ["tests/**"]
